Move filter code to util/slices
