+++ internal/util/slices/filter.go
@@ -1,4 +1,4 @@
-package filter
+package slices
 
 import (
 	"slices"
